requirements.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-filter==24.2
drf-spectacular==0.27.*
mysqlclient==2.2.*
Pillow==10.3.*
python-decouple==3.8
gunicorn==22.0.*
django-storages==1.14.*
celery==5.4.*
redis==5.0.*
geopy==2.4.*
openpyxl==3.1.*
xhtml2pdf==0.2.*
